+++ main.go
@@ -75,6 +75,7 @@ func main() {
 
 	// initialize commands
 	commands.Init_Add_Command(rootCmd, &app)
+	commands.Init_Audit_Command(rootCmd, &app)
 	commands.Init_Build_Command(rootCmd, &app)
 	commands.Init_Checkout_Command(rootCmd, &app)
 	commands.Init_Exec_Command(rootCmd, &app)