+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - feat: `install` command now supports `--tidy` flag to run `tidy` command after successful installation
 - feat: `open alias` command, which opens the URL of an alias from `alias.yaml` file in `$HOME/.gpm/bin` folder
 - feat: `open project` command, which opens the URL of a project  from `projects.yaml` file in `$HOME/.gpm/bin` folder
+- feat: `import aliases` command, which loads aliases from a local or web source and merge them with `alias.yaml` file in `$HOME/.gpm` folder
+- feat: `import projects` command, which loads projects from a local or web source and merge them with `projects.yaml` file in `$HOME/.gpm` folder
 
 ## 0.12.0